#0: Watcher function cleanup
