+++ tt_metal/impl/CMakeLists.txt
@@ -19,6 +19,7 @@ set(IMPL_SRC
 	${CMAKE_CURRENT_SOURCE_DIR}/dispatch/data_collection.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/debug/dprint_server.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/debug/watcher_server.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/debug/watcher_device_reader.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/trace/trace.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/trace/trace_buffer.cpp
 )
